requirements.txt
nextcord
pymysql
python-dotenv
pytz
Pillow
requests
